Calculate tracers, e.g., trophic level

Calculate how a tracer propagates through consumption.

Usage

compute_tracer(
  Q_ij,
  inverse_method = c("Penrose_moore", "Standard"),
  type_i,
  tracer_i = rep(1, nrow(Q_ij))
)

Arguments

Q_ij

Consumption of each prey i by predator j in units biomass.

inverse_method

whether to use pseudoinverse or standard inverse

tracer_i

an indicator matrix specifying the traver value

Details

Trophic level \(y_i\) for each predator \(i\) is defined as:

$$ \mathbf{y = l Q^* + 1} $$

where \(\mathbf{Q*}\) is the proportion consumption for each predator (column) of different prey (rows). We identify primary producers as any taxa with no consumption (a column of 0s), and assign them as the first trophic level.

More generically, a tracer might be used to track movement of biomass through consumption. For example, if we have a tracer \(x_i\) that is 1 for the base of the pelagic food chain, and 0 otherwise, then we can calculate the proportion of pelagic vs. nonpelagic biomass for each taxon:

$$ \mathbf{y = l Q^* + x} $$

This then allows us to separate alternative components of the foodweb.