Detailed control for stanza structure

Define a list of control parameters.

Usage

stanza_settings(
  taxa,
  stanza_groups,
  K,
  d,
  Wmat,
  Amax,
  SpawnX,
  Leading,
  fit_K = c(),
  fit_d = c(),
  fit_phi = vector(),
  Amat = NULL,
  Wmatslope,
  STEPS_PER_YEAR = 1,
  comp_weight = c("multinom", "dir", "dirmult")
)

Arguments

taxa

Character vector of taxa included in model.

stanza_groups

character-vector with names corresponding to taxa and elements specifying the multi-stanza group (i.e., age-structured population) for a given taxa

K

numeric-vector with names matching unique(stanza_groups), providing the von Bertalanffy growth coefficient for length

d

numeric-vector with names matching unique(stanza_groups), providing the von Bertalanffy allometric consumption-at-weight (default is 2/3)

Wmat

numeric-vector with names matching unique(stanza_groups), providing the weight-at-maturity relative to asymptotic weight

Amax

numeric-vector with names matching names(stanza_groups), providing the maximum age (in units years) for a given taxon (and the oldest taxon for a given stanza_group is treated as a plus-group)

SpawnX

numeric-vector with names matching unique(stanza_groups), providing the larval vulnerability (density dependence) parameter

Leading

Boolean vector with names matching names(stanza_groups), with TRUE for the taxon for which scale (B or EE) is specified or estimated, where this is then calculated determinstically for other taxa for a given stanza_group

fit_K

Character-vector listing stanza_groups for which K is estimated

fit_d

Character-vector listing stanza_groups for which d is estimated (note that this currently does not work)

fit_phi

Character-vector listing stanza_groups for which the model should estimate annual recruitment deviations, representing nonconsumptive variation in larval survival (e.g., oceanographic advection)

Amat

numeric-vector with names matching unique(stanza_groups), providing the integer age-at-maturity (in units years)

Wmatslope

numeric-vector with names matching unique(stanza_groups), providing the slope at 0.5 maturity for a logistic maturity-at-weight ogive

STEPS_PER_YEAR

integer number of Euler steps per year for calculating integrating individual weight-at-age

comp_weight

method used for weighting age-composition data

Value

An object (list) of S3-class stanza_settings.